+++ examples/airflow-torch-training/TorchBasicExample.py
