+++ examples/airflow-torch-training/local_run_of_callables.py
@@ -9,7 +9,7 @@
 
 import logging
 
-from callables import (
+from airflow_example_torch_train import (
     access_data_callable,
     bring_up_cluster_callable,
     down_cluster,
