+++ examples/airflow-torch-training/requirements.txt
@@ -1,0 +1,4 @@
+runhouse[aws]
+torch
+torchvision
+airflow
